backend/requirements.txt
Flask
flask-cors
python-dotenv
transformers
torch
moviepy
whisper
deep-translator
google-generativeai
